% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_data.R
\name{append_mock_ids}
\alias{append_mock_ids}
\title{Generate dummy data}
\usage{
append_mock_ids(
  input_path,
  output_path = NA,
  format = "CSV",
  read_func = "read_csv_auto"
)
}
\arguments{
\item{input_path}{character path of input data file}

\item{output_path}{character path of output data file}

\item{format}{character output file format}
}
\value{
character path of output data file
}
\description{
Append fake patient identifiers to the synthetic data.
}
