% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{calc_age}
\alias{calc_age}
\title{Calculate age in years}
\usage{
calc_age(point_time, ref_time)
}
\arguments{
\item{point_time}{Date on which age is to be calculated}

\item{ref_time}{Reference time point, typically date of birth}
}
\value{
String vector of ages, rounded down
}
\description{
Calculate age in years
}
