% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean_names}
\alias{clean_names}
\title{Tidies Names}
\usage{
clean_names(name)
}
\arguments{
\item{name}{String vector of names to be cleaned.}
}
\value{
String vector of names in upper case with white space trimmed.
}
\description{
Replaces multiple white-spaces with single white space and converts characters to upper case.
}
