% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_query.R
\name{configure_connection}
\alias{configure_connection}
\title{Configure database connection}
\usage{
configure_connection(con)
}
\arguments{
\item{con}{DBI::DBIConnection}
}
\value{
DBI::DBIConnection
}
\description{
See: \href{https://duckdb.org/docs/sql/configuration.html}{DuckDB Configuration}
}
