% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_logging.R
\name{configure_logging}
\alias{configure_logging}
\title{Configure logging for the data pipeline.}
\usage{
configure_logging(log_threshold = "INFO", log_dir = NA)
}
\arguments{
\item{log_threshold}{The minimum log level to be recorded. Valid values are "TRACE", "DEBUG", "INFO", "WARN", "ERROR", and "FATAL". The default is "INFO".}

\item{log_dir}{The directory where the log file will be stored. If not provided, the log file will be stored in the temporary directory.}
}
\description{
Configure logging using the \href{https://daroczig.github.io/logger/}{logger} package.
}
