% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_query.R
\name{connect}
\alias{connect}
\title{Connect to database}
\usage{
connect(dbdir = ":memory:", ...)
}
\arguments{
\item{dbdir}{The location of the database, default :memory:}

\item{...}{Arguments to DBI::dbConnect(...)}
}
\value{
Database connection handle
}
\description{
See \href{https://duckdb.org/docs/api/r.html}{DuckDB R API}

We want an in-memory database.
}
