% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_binary.R
\name{convert_json_to_struct}
\alias{convert_json_to_struct}
\title{Convert JSON object to an SQL struct}
\usage{
convert_json_to_struct(data)
}
\arguments{
\item{data}{String. JSON data. The data structure is assumed to be an
object (dictionary).}
}
\description{
Convert JSON object to an SQL struct
}
