% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{convert_str_to_date}
\alias{convert_str_to_date}
\title{Convert date from string to date}
\usage{
convert_str_to_date(date_in, dt_format = "\%Y-\%m-\%d")
}
\description{
Convert date from string to date
}
