% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_binary.R
\name{csv_to_binary}
\alias{csv_to_binary}
\title{Convert CSV data into binary format}
\usage{
csv_to_binary(
  input_dir,
  output_dir,
  data_set_id,
  csv_metadata_path = NA,
  delim = ",",
  header = TRUE,
  temp_directory = NA,
  query_path = NA
)
}
\arguments{
\item{input_dir}{character. Path. The directory that contains the raw data files.}

\item{output_dir}{Character. Path of the directory in which to write the output binary data files.}

\item{data_set_id}{character. Data set identifier e.g. "apc", "op"}

\item{csv_metadata_path}{The path of the CSVW metadata file. Default: inst/extdata/metadata/raw//{data_set_id}.json}

\item{delim}{The character sequence which should separate CSV fields. Default: ","}

\item{header}{Logical. Indicates whether the file includes a header row. Default: TRUE}

\item{temp_directory}{Set the directory to which to write temp files. DuckDB configuration option.}

\item{query_path}{Path of the SQL query that defines the data operation.}
}
\value{
List of paths of the new output files.
}
\description{
The code uses DuckDB to perform file format conversion. This function will build an SQL script that is executed
using DuckDB.

By default, this function will read all the CSV files in the \code{input_dir} and convert them to
\href{https://parquet.apache.org/}{Apache Parquet} format.

The resultant binary data files will be written to \code{output_dir}.

The structure of the CSV file is specified by a \href{https://csvw.org/}{CSVW} document specified by \code{csv_metadata_path}.
This should include a \href{https://specs.frictionlessdata.io/csv-dialect/}{CSV dialect} specification if a non-standard
CSV format is used.

See:
\itemize{
\item \href{https://duckdb.org/docs/api/r}{DuckDB R API}
\item \href{https://dbi.r-dbi.org/}{DBI documentation} and \href{https://duckdb.org/docs/sql/configuration.html}{configuration}
\item DuckDB \href{https://duckdb.org/docs/data/csv/overview.html}{CSV Import}
\item \href{https://csvw.org/}{CSV on the Web}
}
}
