% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_parquet.R
\name{csv_to_parquet}
\alias{csv_to_parquet}
\title{Convert CSV to Parquet format}
\usage{
csv_to_parquet(input_path, output_path)
}
\arguments{
\item{input_path}{String. Source data file location.}

\item{output_path}{String. Target data file location.}
}
\description{
This doesn't convert data types, but assumes every column contains
strings.
}
