% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvw.R
\name{csvw_columns}
\alias{csvw_columns}
\title{Get \href{https://csvw.org/}{CSVW} columns for a table}
\usage{
csvw_columns(csvw_table)
}
\arguments{
\item{csvw_table}{\href{https://w3c.github.io/csvw/syntax/#tables}{CSVW table}}
}
\value{
data.frame \href{https://w3c.github.io/csvw/syntax/#columns}{CSVW columns}
}
\description{
Get \href{https://csvw.org/}{CSVW} columns for a table
}
