% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvw.R
\name{csvw_tables}
\alias{csvw_tables}
\title{Generate CSVW tables}
\usage{
csvw_tables(metadata)
}
\arguments{
\item{csvw}{CSVW document e.g. \code{metadata = jsonlite::fromJSON("my_data_set.json")}}
}
\value{
Iterator of \href{https://w3c.github.io/csvw/syntax/#tables}{CSVW tables}
}
\description{
This is a \href{https://coro.r-lib.org/#generators}{generator} that makes it easy to iterate over the
\href{https://w3c.github.io/csvw/syntax/#tables}{tables} that are listed in a \href{https://csvw.org/}{CSVW document}.
}
