% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvw_to_arrow_schema.R
\name{csvw_to_arrow_schema}
\alias{csvw_to_arrow_schema}
\title{Convert CSVW columns metadata to an Arrow schema}
\usage{
csvw_to_arrow_schema(columns)
}
\description{
\href{https://w3c.github.io/csvw/syntax/}{CSV for the Web}
\href{https://arrow.apache.org/docs/dev/r/reference/Schema.html}{Arrow Schema}
}
