% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_to_dictionary.R
\name{dataframe_to_dictionary}
\alias{dataframe_to_dictionary}
\title{Convert data frame to a dictionary}
\usage{
dataframe_to_dictionary(data_frame, key_col, value_col)
}
\arguments{
\item{data_frame}{data.frame to convert.}

\item{key_col}{The column to use for the dictionary keys}

\item{value_col}{The column to use as the values}
}
\value{
A dictionary.
}
\description{
Converts the first two columns of data frame to a mapping between the values of the two columns.
}
\examples{
data_frame <- data.frame(name = c("Alice", "Bob", "Charlie"), age = c(30, 25, 42))
dictionary <- dataframe_to_dictionary(data_frame, key_col = "name", value_col = "age")
print(dictionary)
}
