% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{date_format}
\alias{date_format}
\title{Is this a valid date}
\usage{
date_format(x)
}
\arguments{
\item{x}{Input object to check}
}
\value{
logical. If TRUE, this is a valid date.
}
\description{
Compare an object's string representation to \href{https://en.wikipedia.org/wiki/ISO_8601#Dates}{ISO 8601 date}.
}
\examples{
date_format("1970-01-01")

}
