% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{date_to_isodate}
\alias{date_to_isodate}
\title{Format date to ISO8601}
\usage{
date_to_isodate(in_date, format_type)
}
\arguments{
\item{in_date}{Input date}

\item{format_type}{Input format, should be one of "yyyy-mm-dd", "yyyymmdd" or
"dd/mm/yyyy"}
}
\value{
String vector of dates formatted as yyyy-mm-dd
}
\description{
Format date to ISO8601
}
\examples{
dates <- c("01/02/1999", "03/04/2001")
date_to_isodate(dates, "dd/mm/yyyy")
dates <- c("19990201", "20010403")
date_to_isodate(dates, "yyyymmdd")
dates <- c("1999-02-01 16:01", "2001-04-03 09:00")
date_to_isodate(dates, "yyyy-mm-dd etc")
}
