% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{datetime_to_isodatetime}
\alias{datetime_to_isodatetime}
\title{Converts datetime to iso datetime.}
\usage{
datetime_to_isodatetime(in_times, format_type)
}
\description{
Converts datetime to iso datetime.
}
