% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tos.R
\name{df2json}
\alias{df2json}
\title{Serialise a data frame to a JSON file.}
\usage{
df2json(df, outputdir, json_file)
}
\arguments{
\item{df}{data.frame Data frame.}

\item{outputdir}{character. Path. Output directory.}

\item{json_file}{character. Path. File name for the JSON file.}
}
\description{
Serialise a data frame to a JSON file.
}
