% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_to_struct.R
\name{dictionary_to_struct}
\alias{dictionary_to_struct}
\title{Convert a dictionary to an \href{https://duckdb.org/docs/sql/data_types/struct.html}{SQL struct}}
\usage{
dictionary_to_struct(dictionary)
}
\arguments{
\item{dictionary}{List. dictionary, key-value pairs.}
}
\value{
SQL struct (text data)
}
\description{
Convert a dictionary to an \href{https://duckdb.org/docs/sql/data_types/struct.html}{SQL struct}
}
