% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata_path.R
\name{extdata_path}
\alias{extdata_path}
\title{Get absolute path for an extdata file in this package
https://r-pkgs.org/data.html#sec-data-extdata}
\usage{
extdata_path(filename = "", mustWork = TRUE)
}
\arguments{
\item{mustWork}{Logical. Raise an error if the file is not found.}

\item{path}{String. Path.}
}
\description{
Get absolute path for an extdata file in this package
https://r-pkgs.org/data.html#sec-data-extdata
}
