% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Extract data from SQL queries}
\usage{
extract(working_dir, clean_dir, sep = .Platform$file.sep)
}
\arguments{
\item{working_dir}{The path of the directory containing the extract definition.}

\item{clean_dir}{The path of the directory containing the clean data to be used in the queries.}
}
\value{
Path of the directory that contains the data extract.
}
\description{
This function extracts data from SQL queries and saves it to Apache Parquet format.

An extract is defined by the following files:
\itemize{
\item Each input SQL file represents an output data table e.g. \code{apc.sql}
\item There must be a \code{README.txt} file in the extract.
}

For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{└───my_extract
    ├──README.txt
    ├──apc.sql
    └──op.sql
}\if{html}{\out{</div>}}

This function will read data from the clean data directory using the supplied SQL queries and write output
data tables to the extract working directory.
}
