% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_ext.R
\name{file_ext}
\alias{file_ext}
\title{Get the file extension}
\usage{
file_ext(path)
}
\arguments{
\item{path}{File path}
}
\value{
File extension of the file \code{path}
}
\description{
Get the file extension
}
