% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_data.R
\name{generate_demographics}
\alias{generate_demographics}
\title{Generate synthetic demographics data and write it to a specified file.}
\usage{
generate_demographics(output_path, n_rows = 1000)
}
\arguments{
\item{output_path}{The path to the file where the generated data will be written.}

\item{n_rows}{The number of rows to generate in the synthetic demographics data (default is 1000).}
}
\description{
This function generates synthetic demographics data by executing a SQL query
template read from a file. The generated data is then written to the specified
output file path.
}
