% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_data.R
\name{generate_mock_demographics_data}
\alias{generate_mock_demographics_data}
\title{Generate synthetic patient demographics data}
\usage{
generate_mock_demographics_data(path, n_rows = 100)
}
\arguments{
\item{path}{character. Output data file path.}

\item{n_rows}{integer number of rows to generate}
}
\value{
character. Output data file path.
}
\description{
Generate synthetic patient demographics data
}
