% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_data.R
\name{generate_synthetic_data}
\alias{generate_synthetic_data}
\title{Generate synthetic data}
\usage{
generate_synthetic_data(query_template_path, output_path = NA)
}
\arguments{
\item{query_template_path}{character. Path of a SQL file.}

\item{output_path}{character. Path of target data file.}
}
\value{
character. Path of destination data file.
}
\description{
Use an SQL query to create some dummy data and save it to a file.
}
