% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_binary.R
\name{get_data_types}
\alias{get_data_types}
\title{Get the data type for each field from the metadata document.}
\usage{
get_data_types(metadata)
}
\value{
Dictionary. Map of field names to data types. For example:
{
"FYEAR": "VARCHAR(4)",
"PARTYEAR": "INT",
"EPIKEY": "VARCHAR(19)",
}
}
\description{
This is the equivalent of this dictionary comprehension in Python:
}
\details{
{field_name: field\link{'data_type'} for field_name, field in metadata.items()}
}
