% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_data.R
\name{get_sample_data}
\alias{get_sample_data}
\title{Generate sample data from NHS Artificial data pilot}
\usage{
get_sample_data(data_urls, number_of_rows = 100)
}
\arguments{
\item{data_urls}{character, set of URIs of CSV data files to download}

\item{number_of_rows}{integer, How many rows to export}
}
\description{
NHS Digital \href{https://digital.nhs.uk/services/artificial-data}{Artificial data pilot}
}
