% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_data.R
\name{get_sample_data_file}
\alias{get_sample_data_file}
\title{Download a sample data file}
\usage{
get_sample_data_file(url, number_of_rows, output_path = NA)
}
\arguments{
\item{url}{character URI of CSV file to download}

\item{number_of_rows}{integer truncate the output data to this number of lines}

\item{output_path}{character path, export the output data to the file}
}
\value{
Character path of output CSV file
}
\description{
Download a sample data file
}
