% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{hes_apc_metadata}
\alias{hes_apc_metadata}
\title{HES Admitted Patient Care (APC) metadata}
\source{
https://digital.nhs.uk/data-and-information/data-tools-and-services/data-services/hospital-episode-statistics/hospital-episode-statistics-data-dictionary
}
\description{
This is extracted from the Technical Output Specification (TOS) for the Hospital Episode Statistics (HES) data.
}
\section{hes_apc_metadata.json}{

Each item is a field in this data set.
}
