% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historic_address.R
\name{historic_address}
\alias{historic_address}
\title{Append historic address}
\usage{
historic_address(input_path, address_path, output_path, date_column)
}
\arguments{
\item{input_path}{character. Path of the source data file.}

\item{address_path}{character. Path of the historic address data file.}

\item{output_path}{character. Path of the destination data file.}

\item{date_column}{character. The name of the field in the source data that contains the date.}
}
\value{
character. Path of output data file.
}
\description{
This function is used to merge the historic address data.
}
