% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage.R
\name{link}
\alias{link}
\title{Link the data set to the reference data.}
\usage{
link(input_path, output_path)
}
\arguments{
\item{input_path}{Source Parquet file path.}

\item{output_path}{Target Parquet file path.}
}
\description{
Link the data set to the reference data.
}
