% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_config.R
\name{load_config}
\alias{load_config}
\title{Load configuration options}
\usage{
load_config(
  data_set_id = NA,
  active_config = NA,
  config_path = NA,
  data_set_config_path = NA
)
}
\arguments{
\item{active_config}{The specific \href{https://rstudio.github.io/config/articles/config.html}{configuration} to use. Default: "default"}

\item{config_path}{Path of the global configuration file. Default: "extdata/config/config.yaml".}

\item{data_set_config_path}{Path of the data set configuration file. Default: "extdata/config/{data_set_id}.yaml".}
}
\description{
Load configuration options
}
