% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tos.R
\name{load_xls}
\alias{load_xls}
\title{Load TOS files}
\usage{
load_xls(xls_file, sheet = "HES APC TOS", sheet_number = NA)
}
\arguments{
\item{xls_file}{str Path File to Excel file from which TOS are to be loaded.}

\item{sheet}{str Name of worksheet to be loaded.}

\item{sheet_number}{int Number of worksheet to be loaded}
}
\description{
Load TOS files
}
