% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{lsoa_format}
\alias{lsoa_format}
\title{Check LSOA string format}
\usage{
lsoa_format(x)
}
\arguments{
\item{x}{character. String to validate.}
}
\value{
logical. Returns TRUE if the string is a valid LSOA code.
}
\description{
An LSOA is a \href{https://webarchive.nationalarchives.gov.uk/ukgwa/20160106001702/https://www.ons.gov.uk/ons/guide-method/geography/beginner-s-guide/census/super-output-areas--soas-/index.html}{Lower Layer Super Output Area} geographical area clustering code.

See:
\itemize{
\item Wikipedia \href{https://en.wikipedia.org/wiki/GSS_coding_system}{GSS coding system}
\item UK ONS Geography \href{https://geoportal.statistics.gov.uk/datasets/ons::lower-layer-super-output-areas-december-2011-names-and-codes-in-england-and-wales-1/explore}{Lower Layer Super Output Areas (December 2011) Names and Codes in England and Wales}
\item UK Data Service \href{https://statistics.ukdataservice.ac.uk/dataset/2011-census-geography-boundaries-lower-layer-super-output-areas-and-data-zones/resource/d65a0201-7240-4d18-8ab2-9d8c1ad830d6}{Lower Super Output Areas and Data Zones borders} contains all possible LSOA values.
}
}
\examples{
lsoa_format("E01014417") # TRUE
lsoa_format("E101447") # FALSE
}
