% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Run workflow}
\usage{
main(data_set_id)
}
\arguments{
\item{data_set_id}{character. Data set identifier e.g. "apc", "yas_epr", "op"}
}
\description{
The function executes the entire data pipeline for the specified data set.
}
