% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parquet.R
\name{parquet_metadata}
\alias{parquet_metadata}
\title{Read \href{https://duckdb.org/docs/data/parquet/metadata.html}{Parquet metadata}}
\usage{
parquet_metadata(path)
}
\arguments{
\item{path}{Character Input data file path.}
}
\value{
data.frame. Column descriptions for the input data file.
}
\description{
Read \href{https://duckdb.org/docs/data/parquet/metadata.html}{Parquet metadata}
}
