% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvw.R
\name{read.csvw}
\alias{read.csvw}
\title{Load a \href{https://csvw.org/}{CSVW} document}
\usage{
read.csvw(path)
}
\arguments{
\item{path}{character. Path of CSVW file}
}
\value{
Dictionary. CSVW metadata
}
\description{
Parse the JSON document that contains the \href{https://csvw.org/}{CSVW} metadata for a data set.
}
