% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{remove_values}
\alias{remove_values}
\title{Removes some values}
\usage{
remove_values(field, values_to_remove, case_sensitive = FALSE)
}
\arguments{
\item{field}{Variable to be cleaned.}

\item{values_to_remove}{values to remove, can be a single value or a list of values.}

\item{case_sensitive}{Boolean whether to make search case-sensitive or not.}
}
\description{
Removes some values
}
