% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{remove_whitespace}
\alias{remove_whitespace}
\title{Remove white spaces from strings}
\usage{
remove_whitespace(x)
}
\arguments{
\item{x}{String vector to tidy}
}
\value{
String vector with carriage returns removed, white space trimmed and missing ("") values as NA.
}
\description{
Remove white spaces from strings
}
