% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{serialise_validation}
\alias{serialise_validation}
\title{Save results summary to a CSV file}
\usage{
serialise_validation(results, path = NA)
}
\arguments{
\item{results}{The output of validate::confront}

\item{path}{character, file path of output CSV file.}
}
\description{
See \href{https://cran.r-project.org/web/packages/validate/vignettes/cookbook.html#11_A_quick_example}{the example}
in the The Data Validation Cookbook.
}
