% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{serialise_validation_plot}
\alias{serialise_validation_plot}
\title{Visualise results}
\usage{
serialise_validation_plot(results, path = NA, ...)
}
\arguments{
\item{results}{The output of validate::confront}

\item{path}{character, file path of output CSV file.}

\item{...}{Arguments for plot(...)}
}
\description{
See the \href{https://cran.r-project.org/web/packages/validate/vignettes/cookbook.html#11_A_quick_example}{Data Validation Cookbook}
}
