% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{split_forenames_surnames}
\alias{split_forenames_surnames}
\title{Split Forenames and Surnames}
\usage{
split_forenames_surnames(names)
}
\arguments{
\item{names}{String vector of names to be split.}
}
\value{
Data Table of forename and surname.
}
\description{
Splits are made based on the last whitespace and so if there are middle names they are included in the Forename
}
