% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_dir.R
\name{temp_dir}
\alias{temp_dir}
\title{Create a temporary directory}
\usage{
temp_dir(check = FALSE)
}
\arguments{
\item{check}{logical indicating if temp_dir() should be checked and recreated if no longer valid.}
}
\description{
R can't handle long path names on Windows, so use a short path.
See \href{https://github.com/CUREd-Plus/cuRed/issues/75}{issue 75}.
}
\details{
This is a wrapper for tempdir(), which is documented here:
https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/tempfile
}
