% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip_files.R
\name{unzip_files}
\alias{unzip_files}
\title{Extract ZIP files from a specified directory}
\usage{
unzip_files(dir_path, glob = "*.zip")
}
\arguments{
\item{dir_path}{The path to the directory containing the ZIP files to extract.}

\item{glob}{A pattern to match the ZIP file names. The default pattern is "*.zip", which matches all ZIP files.}
}
\value{
A vector of paths to the extracted files.
}
\description{
Recursively extracts all ZIP files found in a specified directory.
}
