% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{validate_ae_diagnoses}
\alias{validate_ae_diagnoses}
\title{Tidies A&E Diagnoses}
\usage{
validate_ae_diagnoses(diags)
}
\arguments{
\item{diags}{Vector of diagnoses.}
}
\value{
Vector of diagnoses with values converted to upper case, missing and invalud values as NA.
}
\description{
Tidies A&E Diagnoses
}
