% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{validate_ae_investigations}
\alias{validate_ae_investigations}
\title{Tidies A&E Investigation codes}
\usage{
validate_ae_investigations(invests)
}
\arguments{
\item{invests}{Vector of investigation codes.}
}
\value{
Vector of investigation codes padded with zeros and missing/invalid codes converted to NA.
}
\description{
Ensures A&E Investigation codes are strings of two characters left-padded with two leading zeros.
}
