% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{validate_codes}
\alias{validate_codes}
\title{Tidies Codes}
\usage{
validate_codes(field_in, valid_codes, invalid_code)
}
\arguments{
\item{field_in}{Variable to be validated.}

\item{valid_codes}{List of valid codes.}

\item{invalid_code}{Invalid code to replace values not observed in valid_codes with.}
}
\value{
Vector of codes wher values not in valid_codes are replaced with invalid_code.
}
\description{
Replaces missing and codes that are not valid with the invalid code value.
}
