% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_data}
\alias{validate_data}
\title{Validate a data set}
\usage{
validate_data(data_path, rules_path, output_dir = NA)
}
\arguments{
\item{data_path}{character. Path of input data file.}

\item{rules_path}{character. Validation rules file path.}

\item{output_dir}{character. Path of directory to save results.}
}
\description{
This uses the \href{https://cran.r-project.org/web/packages/validate/}{validate} package.

Please read the \href{https://cran.r-project.org/web/packages/validate/vignettes/cookbook.html}{data validation cookbook}.
}
