% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{validate_date_times}
\alias{validate_date_times}
\title{Tidites Datetimes to POSIX values}
\usage{
validate_date_times(
  datetime_in,
  invalid_code,
  start_date_time,
  finish_date_time,
  missing_date_time
)
}
\arguments{
\item{start_date_time}{The first date from which values are valid, anything before this is converted to NA.}

\item{finish_date_time}{The last date from which values are valid, anything after this is converted to NA.}

\item{missing_date_time}{String representation of date/time that should be considered missing, such values are
converted to NA.}
}
\description{
Converts date-times as strings to elapsed POSIX values. This is required so that elapsed periods can be calculated.
}
