% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{validate_dates}
\alias{validate_dates}
\title{Tidies dates}
\usage{
validate_dates(
  x,
  start_date = "1888-01-01",
  finish_date = "2023-03-31",
  missing_date = "1900-01-01",
  format = "\%Y-\%m-\%d",
  tz = "Europe/London",
  lt = FALSE
)
}
\arguments{
\item{x}{Date variable to tidy.}

\item{start_date}{The first date from which values are valid, anything before this is converted to NA. Default is
"1888-01-01".}

\item{finish_date}{The last date from which values are valid, anything after this is converted to NA. Default is
"2017-03-31".}

\item{format}{Date format, default is \%Y-\%m-\%d}

\item{tz}{Timezone, default is FALSE}

\item{lt}{logical, default FALSE. Whether to return POSIXlt (TRUE) or POSIXct (FALSE).}
}
\value{
Vector of dates tidied.
}
\description{
Replaces "1900-01-01" with NA and anything < "1888-01-01" or > "2023-03-31" as NA.
}
\examples{
x <- c("2001-06-09", "1887-12-31", "2017-04-01", "2016/05/03", NA)
validate_dates(x)
validate_dates(x, start_date = "1887-01-01")
validate_dates(x, finish_date = "2018-03-31")
validate_dates(x, format = "\%Y/\%m/\%d")
}
