% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{validate_digits}
\alias{validate_digits}
\title{Tidies numerical values to be strings padded with 0}
\usage{
validate_digits(field_in, len, valid_codes, invalid_code)
}
\description{
Pads numerical values with zeros, values NOT in valid_codes are replaced with invalid_code.
}
