% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{validate_gp_prac}
\alias{validate_gp_prac}
\title{Tidies GP Practice code to upper-case.}
\usage{
validate_gp_prac(field_in, invalid_code)
}
\arguments{
\item{invalid_code}{Code to indicate invalid or NA value.}
}
\description{
Converts GP Practice codes to upper case and replaces those that do not conform or are missing with an invalid code.
}
