% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{validate_icd10}
\alias{validate_icd10}
\title{Tidies ICD10 codes}
\usage{
validate_icd10(diags_wide, apc_data = TRUE)
}
\arguments{
\item{diags_wide}{Vector of ICD10 codes, can contain multiple codes per cell, as these will be split.}

\item{apc_data}{APC data}
}
\value{
Data table of multiple ICD10 diagnoses.
}
\description{
Tidies ICD10 codes
}
\examples{
icd10 <- c("A00.0", "I60.2", "I60.10", NA)
validate_icd10(icd10, apc_data = TRUE)
}
