% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{validate_nhs_number}
\alias{validate_nhs_number}
\title{Tidies NHS Numbers}
\usage{
validate_nhs_number(nhs_number)
}
\arguments{
\item{nhs_number}{String vector of nhs numbers.}
}
\value{
String vector of ten digits. Values that do not conform are converted to NA.
}
\description{
Tidies NHS Numbers
}
