% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{validate_numeric}
\alias{validate_numeric}
\title{Validates numerical values}
\usage{
validate_numeric(x, len)
}
\arguments{
\item{x}{Vector of numerical values.}

\item{len}{Length entries are meant to be.}
}
\value{
String vector of numbers left-padded with zero. Missing values are '9' repeated the value of 'len'.
}
\description{
Validates numerical values
}
