% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{validate_opcs_ops}
\alias{validate_opcs_ops}
\title{Tidies OPCSOps codes}
\usage{
validate_opcs_ops(ops_wide, invalid_date_code)
}
\arguments{
\item{ops_wide}{OPCS Data in wide format.}

\item{invalid_date_code}{Code to use for invalid dates.}
}
\value{
Data table of tidied formats in upper case with missing dates replaced.
}
\description{
Tidies OPCSOps codes
}
