% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{validate_postcode}
\alias{validate_postcode}
\title{Tidies Postcode}
\usage{
validate_postcode(postcode)
}
\arguments{
\item{postcode}{String vector of postcodes.}
}
\value{
String vector of tidied postcodes.
}
\description{
Formats postcode correctly ensuring all characters are uppercase and incorrect values as NA
}
