% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{validate_procodet}
\alias{validate_procodet}
\title{Tidies Procodet}
\usage{
validate_procodet(procodet_in, procodet_sender)
}
\arguments{
\item{prodocdet_in}{Prodocdet input}

\item{prodocdet_sender}{Value to replace missing observations or when the length of the code is neither 3 nor 5.}
}
\value{
Tidied prodocdet of length 3 characters with no missing values.
}
\description{
Tidies Procodet
}
