% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{validate_refer_org}
\alias{validate_refer_org}
\title{Tidies Referer organisation code to upper-case.}
\usage{
validate_refer_org(field_in, invalid_code)
}
\description{
Converts Referer organisation codes to upper case and replaces those that do not conform or are missing with an
invalid code.
}
