% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{verify}
\alias{verify}
\title{Verify the schema of a data file using a rule set.}
\usage{
verify(file_path, rules)
}
\arguments{
\item{file_path}{String. Input data file path.}

\item{rules}{String. YAML file path, data validation rules.}
}
\value{
Data validation results, one row per rule.
}
\description{
Uses the \href{https://davzim.github.io/dataverifyr/index.html}{dataverifyr package}.
}
