% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_file.R
\name{write_file}
\alias{write_file}
\title{Serialise text data to a file}
\usage{
write_file(file_path, data)
}
\arguments{
\item{file_path}{Target text file path.}

\item{string}{Text data to write.}
}
\description{
Serialise text data to a file
}
