% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type.R
\name{xml_schema_to_arrow}
\alias{xml_schema_to_arrow}
\title{Convert XML schema data type to an Arrow schema field}
\usage{
xml_schema_to_arrow(xml_data_type)
}
\arguments{
\item{xml_data_type}{The XML data type to convert.}
}
\value{
The arrow Field
}
\description{
Converts an XML data type to a field type for an Arrow schema.

\href{https://w3c.github.io/csvw/syntax/#dfn-datatype}{XML Schema data type} →
\href{https://arrow.apache.org/docs/r/reference/data-type.html}{Arrow data type}
}
