% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type.R
\name{xml_schema_to_sql_data_type}
\alias{xml_schema_to_sql_data_type}
\title{Convert XML schema data type to an SQL data type}
\usage{
xml_schema_to_sql_data_type(xml_data_type)
}
\arguments{
\item{xml_data_type}{The XML data type to convert.}
}
\value{
The SQL data type.
}
\description{
Converts an XML data type to a SQL data type.

See:
\itemize{
\item \href{https://w3c.github.io/csvw/syntax/#dfn-datatype}{XML Schema data types}
\item \href{https://duckdb.org/docs/sql/data_types/overview}{DuckDB SQL data types}
}
}
