% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type.R
\name{yas_type_to_data_type}
\alias{yas_type_to_data_type}
\title{Convert YAS data type to an SQL data type}
\usage{
yas_type_to_data_type(format_str)
}
\arguments{
\item{format_str}{character. Data type e.g. "boolean", "categorical", "numerical"}
}
\value{
character. \href{https://duckdb.org/docs/sql/data_types/overview.html}{DuckDB data types}
}
\description{
Decide what data type is appropriate to use to store this data type
in an SQL database.
}
