% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_data.R
\name{append_mock_ids}
\alias{append_mock_ids}
\title{Append generated dummy data in CSV format.}
\usage{
append_mock_ids(input_path, output_path)
}
\arguments{
\item{input_path}{character path of input CSV data file (or files, e.g. "*.csv")}

\item{output_path}{character path of output CSV data file}
}
\value{
character path of output data file
}
\description{
Append fake patient identifiers to the synthetic data.
}
