% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{artificial_hes_apc_0102}
\alias{artificial_hes_apc_0102}
\title{HES Admitted Patient Care artificial data set}
\source{
https://digital.nhs.uk/services/artificial-data
}
\description{
\href{https://digital.nhs.uk/services/artificial-data}{NHS Artificial data pilot}
}
\details{
The documentation and the artificial data files are © Crown copyright and available under the terms of the
\href{https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/}{Open Government 3.0} licence.
}
\section{artificial_hes_apc_0102.csv}{

}

