% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_query.R
\name{connect}
\alias{connect}
\title{Connect to database
https://duckdb.org/docs/api/r.html
https://dbi.r-dbi.org/reference/dbconnect}
\usage{
connect(dbdir = ":memory:", ..)
}
\arguments{
\item{dbdir}{The location of the database, default :memory:}
}
\value{
Database connection handle
}
\description{
Connect to database
https://duckdb.org/docs/api/r.html
https://dbi.r-dbi.org/reference/dbconnect
}
