% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{convert_str_to_date}
\alias{convert_str_to_date}
\title{Convert date from string to date}
\usage{
convert_str_to_date(date, dt_format = "\%Y-\%m-\%d")
}
\arguments{
\item{date}{Vector of dates as strings to be converted.}

\item{dt_format}{Format of date strings.}
}
\value{
Vector of dates as elapsed dates.
}
\description{
Convert date from string to date
}
\examples{
dates_YYYY_MM_DD <- c("1999-10-06", "2000-01-01", "1900-01-01")
cuRed::convert_str_to_date(dates_YYYY_MM_DD)
dates_DD_MM_YYYY <- c("06-10-1999", "01-01-2000", "01-01-1900")
cuRed::convert_str_to_date(dates_DD_MM_YYYY, dt_format = "\%d-\%m-\%Y")
dates_DD_MM_YYYY <- c("06/10/1999", "01/01/2000", "01/01/1900")
cuRed::convert_str_to_date(dates_DD_MM_YYYY, dt_format = "\%d/\%m/\%Y")
}
