% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_rows.R
\name{count_rows}
\alias{count_rows}
\title{Count the number of rows in the input data set.}
\usage{
count_rows(path)
}
\arguments{
\item{path}{String. File path of a Parquet file.}
}
\description{
Count the number of rows in the input data set.
}
