% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_binary.R
\name{csv_to_binary}
\alias{csv_to_binary}
\title{Convert CSV data into binary format}
\usage{
csv_to_binary(input_dir, output_dir, data_set_id)
}
\arguments{
\item{input_dir}{character. Path. The directory that contains the raw data files.}

\item{output_dir}{Character. Path of the directory in which to write the output binary data files.}

\item{data_set_id}{character. Data set identifier e.g. "apc", "op"}
}
\value{
List of paths of the new output files.
}
\description{
The code uses DuckDB to perform file format conversion.

By default, this function will read all the CSV files in the \code{input_dir} and convert them to
\href{https://parquet.apache.org/}{Apache Parquet} format.

The resultant binary data files will be written to \code{output_dir}.
}
