% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_binary.R
\name{csv_to_binary}
\alias{csv_to_binary}
\title{Use Duck DB to perform file format conversion.}
\usage{
csv_to_binary(raw_data_dir, output_data_dir, metadata)
}
\arguments{
\item{raw_data_dir}{String. Path. The directory that contains the raw data files.}

\item{output_data_dir}{String. Path. The directory the output data file(s) should be written to.}

\item{metadata}{List. Dictionary containing the column definitions.}
}
\value{
String. Path. The path of the output data file.
}
\description{
DuckDB R API
https://duckdb.org/docs/archive/0.8.1/api/r
}
\details{
DBI documentation
https://dbi.r-dbi.org/

Example: "Quickly Convert CSV to Parquet with DuckDB"
https://rmoff.net/2023/03/14/quickly-convert-csv-to-parquet-with-duckdb
}
