% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{datetime_to_isodatetime}
\alias{datetime_to_isodatetime}
\title{Converts datetime to iso datetime.}
\usage{
datetime_to_isodatetime(x, format_type)
}
\arguments{
\item{x}{String vector of date-times to be converted.}

\item{format_type}{Format of date-times.}
}
\value{
Vector of dates in ISO-8601 format.
}
\description{
Converts datetime to iso datetime.
}
\examples{
date_time <- c("2018-11-14 09:16:43", "2017-10-28 20:01:59", NA)
datetime_to_isodatetime(date_time, format_type = "YYYY-MM-DD HH:MM:SS")
}
