% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tos.R
\name{df2json}
\alias{df2json}
\title{write dataframe to json}
\usage{
df2json(df, outputdir, json_file)
}
\arguments{
\item{df}{dataframe(tibble)}

\item{outputdir}{str}

\item{json_file}{str}
}
\description{
write dataframe to json
}
