% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata_path.R
\name{extdata_path}
\alias{extdata_path}
\title{Get an extdata file path}
\usage{
extdata_path(path = "", mustWork = TRUE)
}
\arguments{
\item{path}{Character. The relative path of file within inst/extdata.}

\item{mustWork}{Logical. Raise an error if the file is not found}
}
\value{
character. The file path
}
\description{
This function builds the absolute path for an extdata file in this package.

See: \href{https://r-pkgs.org/data.html#sec-data-extdata}{Raw data file} in \emph{R Packages} by H. Wickham.
}
