% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata_path.R
\name{extdata_path}
\alias{extdata_path}
\title{Get the path of an external data file}
\usage{
extdata_path(path = "", mustWork = TRUE)
}
\arguments{
\item{path}{Character. Path of file within inst/extdata}

\item{mustWork}{Logical. Raise an error if the file is not found}
}
\value{
The absolute file path of the file.
}
\description{
Build the absolute path of an external data (\code{extdata}) file in this package.

This files are to be found in the \code{inst/extdata} directory.

See: \href{https://r-pkgs.org/data.html#sec-data-extdata}{external data}
}
