% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_binary.R
\name{format_to_data_type}
\alias{format_to_data_type}
\title{Convert TOS format to an SQL data type}
\usage{
format_to_data_type(format_str)
}
\value{
String. SQL data type.
}
\description{
See: NHS Digital "Constructing submission files" for the data formats from
the NHS Data Model and Dictionary.
}
\details{
\href{https://duckdb.org/docs/sql/data_types/overview.html}{DuckDB data types}
}
