% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_binary.R
\name{get_data_types}
\alias{get_data_types}
\title{Get data types}
\usage{
get_data_types(metadata, data_set_id)
}
\arguments{
\item{metadata}{List of field objects.}

\item{data_set_id}{Data set identifier}
}
\value{
Dictionary. Map of field names -> data types.
}
\description{
Get the data type for each field from the metadata document.
}
