% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_binary.R
\name{get_data_types}
\alias{get_data_types}
\title{Get data types}
\usage{
get_data_types(metadata)
}
\arguments{
\item{metadata}{Nested dictionary. The keys are the field names.}
}
\value{
Dictionary. Map of field names to data types.
}
\description{
Get the data type for each field from the metadata document.
}
