% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_query.R
\name{get_query}
\alias{get_query}
\title{Get SQL query result using the \href{https://duckdb.org/docs/api/r.html}{DuckDB R API}}
\usage{
get_query(query, read_only = FALSE)
}
\arguments{
\item{query}{character. SQL query}

\item{read_only}{logical. If true, use READ_ONLY access mode}
}
\value{
Data frame containing query results.
}
\description{
Get SQL query result using the \href{https://duckdb.org/docs/api/r.html}{DuckDB R API}
}
