% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{link}
\alias{link}
\title{Link the data set to the reference data.}
\usage{
link(input_path, output_path, patient_path, demographics_path, deaths_path)
}
\arguments{
\item{input_path}{Source Parquet file path.}

\item{output_path}{Target Parquet file path.}

\item{patient_path}{String. Path to the patient ID bridge file.}
}
\description{
Link the data set to the reference data.
}
