% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{link}
\alias{link}
\title{Link the data set to the reference data.}
\usage{
link(data_set_id, input_path, output_path, patient_path, demographics_path)
}
\arguments{
\item{data_set_id}{character. Data set identifier e.g. "apc", "ae", "op"}

\item{input_path}{Path of the source data file.}

\item{output_path}{character. Path of the merged data file.}

\item{patient_path}{character. Path of the patient ID bridge file.}

\item{demographics_path}{character. Path of the demographics data file.}
}
\value{
character. Path of output file.
}
\description{
Link the data set to the reference data.
}
