% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Run workflow}
\usage{
main(data_sets_path = NA)
}
\arguments{
\item{data_sets_path}{String. Path. The path of the configuration file. Defaults to the one included in the library.}
}
\description{
\code{main} executes the entire data pipeline for all data sets.
}
\details{
TODO
}
