% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Run workflow}
\usage{
main(data_set_id, config_active = NA)
}
\arguments{
\item{data_set_id}{Character. Data set identifier e.g. "apc", "yas_epr", "op"}

\item{config_active}{Character. The specific \href{https://rstudio.github.io/config/articles/config.html}{configuration} to use.}
}
\description{
The function executes the entire data pipeline for the specified data set.
}
