% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Run workflow}
\usage{
main(root_directory)
}
\arguments{
\item{root_directory}{String. Path. The directory that contains all the working directories.}
}
\description{
\code{main} executes the entire data pipeline for all data sets.
}
\details{
TODO
}
