% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parquet.R
\name{parquet_metadata}
\alias{parquet_metadata}
\title{Read Parquet metadata
https://duckdb.org/docs/data/parquet/metadata.html}
\usage{
parquet_metadata(file_path)
}
\arguments{
\item{file_path}{String. Input data file path.}
}
\value{
Table of column descriptions from the input data file.
}
\description{
Read Parquet metadata
https://duckdb.org/docs/data/parquet/metadata.html
}
