% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tos.R
\name{parse_tos}
\alias{parse_tos}
\title{Parse Technical Output Specification (TOS) metadata file.}
\usage{
parse_tos(metadata_path)
}
\arguments{
\item{metadata_path}{String path. The metadata file i.e. TOS spreadsheet file path.}
}
\description{
Parse Technical Output Specification (TOS) metadata file.
}
