% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tos.R
\name{parse_tos}
\alias{parse_tos}
\title{Parse Technical Output Specification (TOS) metadata file.}
\usage{
parse_tos(xls_file, sheet)
}
\arguments{
\item{xls_file}{character. Path File to Excel file from which TOS are to be loaded.}

\item{sheet}{character. Name of worksheet to be loaded.}
}
\value{
data.frame TOS metadata
}
\description{
Parse Technical Output Specification (TOS) metadata file.
}
