% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parquet.R
\name{read_parquet}
\alias{read_parquet}
\title{Load an Apache Parquet file into an R data frame.}
\usage{
read_parquet(path)
}
\arguments{
\item{path}{Character. Source data file}
}
\value{
data.frame Data file contents.
}
\description{
This function uses the DuckDB \href{https://duckdb.org/docs/data/parquet/overview#read_parquet-function}{read_parquet()}
function.
}
