% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{remove_values}
\alias{remove_values}
\title{Removes some values}
\usage{
remove_values(x, values_to_remove, case_sensitive = FALSE)
}
\arguments{
\item{x}{Variable to be cleaned.}

\item{values_to_remove}{values to remove, can be a single value or a list of values.}

\item{case_sensitive}{Boolean whether to make search case-sensitive or not.}
}
\value{
Vector with values_to_remove converted to NA.
}
\description{
Removes some values
}
\examples{
x <- c("ab", "Cd", "eF", "gh", "IJ")
remove_case_insensitive <- c("ef", "ij")
remove_values(x, remove_case_insensitive, case_sensitive = FALSE)
remove_case_sensitive <- c("ef", "IJ")
remove_values(x, remove_case_sensitive, case_sensitive = FALSE)
}
