% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_query.R
\name{run_query}
\alias{run_query}
\title{Run a DuckDB query}
\usage{
run_query(query)
}
\arguments{
\item{query}{Structured Query Language (SQL) https://duckdb.org/docs/sql/introduction}
}
\value{
The number of rows affected by the query.
}
\description{
Run a DuckDB query
}
