% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_workflow.R
\name{run_workflow}
\alias{run_workflow}
\title{Run a data processing workflow for a single data set.}
\usage{
run_workflow(
  data_set_id,
  raw_data_dir,
  metadata_path,
  sheet,
  staging_dir,
  patient_path,
  demographics_path,
  address_path
)
}
\arguments{
\item{data_set_id}{character. Data set identifier e.g. "apc" or "op"}

\item{raw_data_dir}{character. The directory that contains the raw data for this data set.}

\item{metadata_path}{character. Path of the technical output specification (TOS) workbook file path.}

\item{sheet}{character. Name of the sheet (tab) in the TOS workbook}

\item{staging_dir}{character. Path of the directory to store working data files.}

\item{patient_path}{character. Path of the patient identifier bridge data file.}

\item{demographics_path}{character. Path of the demographics file.}

\item{address_path}{character. Path of the historic address data file.}
}
\description{
\code{run_workflow} executes the data pipeline for a single data set.
}
\details{
TODO
}
