% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_dir.R
\name{temp_dir}
\alias{temp_dir}
\title{Create a temporary directory with a random path.}
\usage{
temp_dir(check = FALSE)
}
\arguments{
\item{check}{logical indicating if temp_dir() should be checked and recreated if no longer valid.}
}
\description{
The directory will be deleted when the session ends, as tempdir() creates a per-session directory path.
}
\details{
On Windows, the temporary directory is determined by the \code{TEMP} environment variable.

R can't handle long path names on Windows, so use a short path.
See \href{https://github.com/CUREd-Plus/cuRed/issues/75}{issue 75}.

This is a wrapper for tempdir(), which is documented here:
https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/tempfile
}
