% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{time_to_isotime}
\alias{time_to_isotime}
\title{Convert times to standardised strings.}
\usage{
time_to_isotime(x, format_type)
}
\arguments{
\item{x}{Vector of times to convert.}

\item{format_type}{Format times take}
}
\value{
Vector of standardised times as string formatted as HH:MM:SS
}
\description{
Convert times to standardised strings.
}
\examples{
times <- c("91643", "200159", NA)
time_to_isotime(times, format_type = "HHMMSS")
times <- c("2018-11-14 09:16:43", "2017-10-28 20:01:59", NA)
time_to_isotime(times, format_type = "yyyy-mm-dd HH:MM:SS")
}
