% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate input raw data}
\usage{
validate(data_path, rules_path)
}
\arguments{
\item{data_path}{String. Source data file path.}

\item{rules_path}{String. Data validation rules file path.}
}
\description{
This uses the \href{https://davzim.github.io/dataverifyr/}{Data vertifyer package}.
}
