% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate a data set}
\usage{
validate(data_dir, metadata)
}
\arguments{
\item{data_dir}{String path. The path of the directory containing the data.}

\item{metadata}{The validation rules}
}
\description{
Stop and raise an error if any of the rules aren't met.
}
