% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{validate}
\alias{validate}
\title{Tidies dates}
\usage{
validate(field_in, start_date, finish_date)
}
\arguments{
\item{field_in}{Date variable to tidy.}

\item{start_date}{The first date from which values are valid, anything before this is converted to NA.}

\item{finish_date}{The last date from which values are valid, anything after this is converted to NA.}
}
\value{
Vector of dates tidied.
}
\description{
Replaces "1900-01-01" with NA and anything < "188-01-01" or > "2017-03-31" as NA.
}
