% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{validate_ae_times}
\alias{validate_ae_times}
\title{Tidies A&E Times}
\usage{
validate_ae_times(x, invalid_code)
}
\arguments{
\item{x}{String vector of times.}

\item{invalid_code}{Code to use for invalid values.}
}
\value{
Vector of times correctly formatted.
}
\description{
Tidies A&E times, converting 2400 to 0000 and replacing invalid values with the specified code.
}
\examples{
ae_times <- c("0113", "3106", NA, "2400")
validate_ae_times(ae_times)
}
