% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{validate_ae_treatments}
\alias{validate_ae_treatments}
\title{Validate A&E Treatments}
\usage{
validate_ae_treatments(x, width = 4, side = "left", pad = "0")
}
\arguments{
\item{x}{Treatment code to tidy and validate}

\item{width}{Width (total number of characters) to pad string. Default is 2}

\item{side}{Side to pad string, default is "left".}

\item{pad}{String to pad with.}
}
\value{
Vector of treatment codes padded with zeros and missing/invalid codes converted to NA.
}
\description{
Tidies treatment codes to be strings, left padded with zeros, if not possible returns NA
}
\details{
Values not conforming are replaced with NA.
}
\examples{
treatment_codes <- c("123", "45", "6")
validate_ae_treatments(treatment_codes)
}
