% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{validate_sitetret}
\alias{validate_sitetret}
\title{Tidies Site Treatment codes to upper-case}
\usage{
validate_sitetret(sitetret_in, procodet)
}
\description{
Converts Site Treatment codes to upper case and replaces those that do not conform or are missing with an invalid
code.
}
